+++ index.js
@@ -4,7 +4,6 @@ module.exports = {
 		'plugin:you-dont-need-lodash-underscore/all-warn',
 	],
 	plugins: [
-		'dependencies',
 		'jsdoc',
 		'mocha',
 		'node',
@@ -79,19 +78,6 @@ module.exports = {
 			'all',
 		],
 
-		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciescase-sensitive
-		'dependencies/case-sensitive': 1,
-
-		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciesno-cycles
-		'dependencies/no-cycles': 1,
-
-		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciesno-unresolved
-		// Duplicates node/no-missing-require
-		'dependencies/no-unresolved': 0,
-
-		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciesrequire-json-ext
-		'dependencies/require-json-ext': 1,
-
 		// http://eslint.org/docs/rules/dot-notation
 		'dot-notation': [
 			'warn',
@@ -165,7 +151,7 @@ module.exports = {
 		'jsdoc/require-hyphen-before-param-description': 0,
 
 		// https://eslint.org/docs/rules/keyword-spacing
-		'keyword-spacing': 0,
+		'keyword-spacing': 1,
 
 		// eslint.org/docs/rules/key-spacing
 		'key-spacing': [
@@ -452,8 +438,7 @@ module.exports = {
 		// http://eslint.org/docs/rules/prefer-const
 		// https://github.com/airbnb/javascript#variables--const
 		// If a variable is never reassigned, using the const declaration is better.
-		// TODO: Turned off for now; revisit.
-		'prefer-const': 'off',
+		'prefer-const': 1,
 
 		// http://eslint.org/docs/rules/prefer-promise-reject-errors
 		'prefer-promise-reject-errors': 'error',