+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG.md
 
+## v2.1.0 - 2019.10.25
+
+- Added [`keyword-spacing`](https://eslint.org/docs/rules/keyword-spacing)
+- Added [`prefer-const`](https://eslint.org/docs/rules/prefer-const)
+- Removed [`eslint-plugin-dependencies`](https://www.npmjs.com/package/eslint-plugin-dependencies) ([abandoned](https://github.com/zertosh/eslint-plugin-dependencies/issues/23))
+- Package updates
+
 ## v2.0.1 - 2019.07.01
 
 - Updated `ecmaVersion` to `2017`